changing everything to int

+++ src/Character/Berserker.h
@@ -10,13 +10,13 @@ private:
     const int MAXRAGE = 100;
     int rage = 0;
 public:
-    Berserker(const string&, double, double);
+    Berserker(const string&, int, int);
     int getStat() const override;
     void setStat(int) override;
     int specialAttack(Character* target) override;
     void reset() override;
     string getInfo() override;
-    void takeDamage(double damage) override; // rage increases when damage is taken
+    void takeDamage(int damage) override; // rage increases when damage is taken
 };
 
 #endif
